Adding front-end implementations for fetching game details
game-details.html:
+ Newly added file.

plugins.js:
+ Added gameDetailToCard() which takes in the details for a game as
parameters then outputs it to a styled <ul>. This can be re-used whether
it's for a single game or can be called multiple times when listing
details for all games.

+ Added executeAfterFetchingGameDetails() similarly to the other AJAX
implementations. It takes an asin, a boolean represented as an int, and
a callback function which will be passed in all the fetched details.