Implemented front- & back-end for deleting developers
DeleteDevelopers.cgi, DeleteDevelopers.py:
+ Newly implemented back-end implementations

manage-devs.html:
+ Newly implemented front-end

fresh.sql:
+ Added sp_DeleteDev which goes through the nested developers table in
games2 deleting anywhere the dev_id shows up, then finally deletes it
from the developers table.

plugins.js:
+ Added redirectToDashobardAfterTimeout()
+ Modified the success check in executeAfterFetchingAllDevelopers to not
check if the devs length is > 0. This is because it is possible for
there to not be any devs in the system, so we don't want to display an
error if that's true.